% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watermark.R
\name{add_details}
\alias{add_details}
\title{Hide in infomation in an image}
\usage{
add_details(p, path)
}
\arguments{
\item{p}{magick-image where details are to be hidden}

\item{path}{character string of file path containing original image}
}
\value{
returns a copy of p with the details imbedded into the image
}
\description{
This function is used by watermark to hide details of the original image and the creation of the watermarked version into the new file.
The purpose is to allow this to be extracted later for replacing images in office documents where they are stripped of structure and filename.
}
