% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASlashB.R
\name{\%a/b\%}
\alias{\%a/b\%}
\title{Returns elements of vector \eqn{x} that are not contained within vector \eqn{y}.}
\usage{
x \%a/b\% y
}
\arguments{
\item{x}{vector with values to be matched.}

\item{y}{vector with values to match against vector \eqn{x}.}
}
\value{
A vector of all values contained in vector \eqn{x} that could not be located in vector \eqn{y}.
}
\description{
Returns elements of vector \eqn{x} that are not contained within vector \eqn{y}.
}
\examples{
union_jack = c("red","white","blue")
rainbow = c("red","orange","yellow","green","blue","indigo","violet")
union_jack \%a/b\% rainbow # gives "white".
# i.e. the only colour in the Union Jack that is not contained in the rainbow is "white".
}
\author{
Will Haese-Hill
}
