% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAST1-package.R
\docType{package}
\name{BAST1}
\alias{BAST1}
\alias{BAST1-package}
\title{BAST1: A package containing BAST R functions.}
\description{
The BAST1 package provides four categories of important functions:
Data building, ....
}
\section{Data building functions}{

The Data building functions ...
}

