% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_stats.R
\name{bin_stats}
\alias{bin_stats}
\title{Split data into bins (usually time bins), then calculate percentiles of data in each bin.}
\usage{
bin_stats(data, bins, dv_col = "DV", ind_var_col = "TIME", pct = c(10, 50,
  90))
}
\arguments{
\item{data}{data frame. Suitably filtered and containing the data to analyse.}

\item{bins}{numeric vector giving the cut off values for the time bins (binning doesn't have to based on time - see ind_var_col).}

\item{dv_col}{column name (default value 'DV') in \eqn{data} containing the experimental data (default value is 'DV')}

\item{ind_var_col}{column name (default value 'TIME') in \eqn{data} containing the variable to use for binning.}

\item{pct}{numeric vector of percentiles (default value c(10,50,90)) to calculate for the dv_col data within each bin.}
}
\value{
A matrix object with one column for each bin, and the following rows of summary statistics for the dv_col data within each bin
(see sum_stats function documentation for descriptions):
\itemize{
   \item min_ind_var_col
   \item max_ind_var_col
   \item n
   \item min
   \item max
   \item median
   \item arith_mean
   \item arith_std_dev
   \item arith_CV
   \item geo_mean
   \item geo_std_dev
   \item geo_CV
   \item requested percentiles (pct)
 }
}
\description{
Essentially the part of VPC that deals with observed data. Good for finding trends in large data sets.
}
\note{
For a bin with limits (LOWER,UPPER), an observation will go into that bin if it is >=LOWER and <UPPER.
}
\author{
Rupert Austin
}
