% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_conc.R
\name{calc_conc}
\alias{calc_conc}
\title{Calculates the molar concentration for a given compound's SI concentration.}
\usage{
calc_conc(conc = 1, MW, SIu, Mu)
}
\arguments{
\item{conc}{numeric. Observed SI concentration of compound.}

\item{MW}{numeric. Molecular weight of compound ( D (g/mol)).}

\item{SIu}{character. SI Units of observed concentration: 'm/V'. Options restricted to:
\itemize{
 \item m = g , mg, ug, ng, pg
 \item V = L , mL, uL, nL, pL
}}

\item{Mu}{character. Desired molar units of concentration: 'M', 'mM', 'uM', 'nM', 'pM'}
}
\value{
numeric. The molar concentration of the compound, with unit \eqn{Mu}.
}
\description{
Calculates the molar concentration for a given compound's SI concentration.
}
\details{
Utilises the formula \deqn{C = ( m / V )( 1 / MW ),}where \eqn{C} is concentration in desired molar units.

Tips:
\itemize{
 \item If SI units given for \eqn{Mu}, and \eqn{MW = 1}, function will perform a simple SI unit conversion.
 \item Set \eqn{conc = 1} (default) if you simply want to return the unit conversion factor.
}
}
\examples{
\donttest{
Observed SI concentration of compound 1000 ug/L.
Compound has molecular weight 494.56 D (g/mol).
Desired units for molar concentration: nM.

new_conc <- calc_conc(conc = 1000, MW = 494.56, SIu = 'ug/L', Mu = 'nM')

new_conc = 2022, so 1000 ug/L of this particular compound is equal to ~2022 nM.
}
}
\author{
Will Haese-Hill
}
