% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_NPC.R
\name{calc_NPC}
\alias{calc_NPC}
\title{Calculate and plot NPC.}
\usage{
calc_NPC(exp_data, sim_data, pct = c(10, 20, 30, 40, 50, 60, 70, 75, 80, 85,
  90), nreps, dv_col = "DV", full = TRUE, save_path = "",
  ind_var_col = "TAD", bins = FALSE, name_stub = "")
}
\arguments{
\item{exp_data}{data frame containing experimental data.}

\item{sim_data}{data frame containing all simulated data. Should have the same data points as exp_data, simply multiple simulations of them.}

\item{pct}{vector of the sizes (in \%) of the prediction intervals to be tested.}

\item{nreps}{numeric. The number of simulations performed.}

\item{dv_col}{name of column containing DV, for both exp_data and sim_data.}

\item{full}{TRUE (default) or FALSE. If TRUE, then the function returns maximum output and plots (see Details).}

\item{save_path}{path for folder to save plots.}

\item{ind_var_col}{name of column for independent variable for normalised NPC plots (normally TIME or TAD), for both exp_data and sim_data.}

\item{bins}{numeric vector defining bins for the normalised NPC plots. If set to FALSE then normalised NPC plots not generated.}

\item{name_stub}{string to be added to the file names of the 3 plots that are saved by the function.}
}
\value{
A series of NPC plots, as well as a list or vector of information (see Details).
}
\description{
Function for performing a Numeric Predictive Check (NPC).
}
\details{
If full==FALSE: \cr
         This function will return a named numeric vector, the same length as pct indicating what \% of the experimental
         data lies within each prediction interval tested.

         If full==TRUE: \cr
         This function will return a list with 8 elements:
         \itemize{
           \item $exp_count
           \item $df_NPC
           \item $pct_within_interval
           \item $pct_above_interval
           \item $pct_below_interval
           \item $n_within_interval
           \item $n_above_interval
           \item $n_below_interval
         }
         $exp_count is simply the number of experimental data points. \cr
         $df_NPC is a copy of exp_data with some columns added: 5th, 50th, 95th percentiles of simulated data; normalised_NPC, defined as
         \deqn{normalised_NPC = (DV - median of simulations) / abs(5th OR 95th percentile of simulations - median of simulations),}
         where 95th percentile is used if \eqn{DV > median} and 5th percentile is used if \eqn{DV < median}. \cr
         If \eqn{normalised_NPC = 0}, then observation = median of simulations. \cr
         If \eqn{normalised_NPC = +1}, then observation = 95th percentile of simulations. \cr
         If \eqn{normalised_NPC = -1}, then observation = 5th percentile of simulations. \cr
         Note that \deqn{$pct_within_interval + $pct_above_interval + $pct_below_interval = 100.}
}
\note{
Deciding if a number is within the prediction interval is inclusive at both ends. i.e. if the number is equal to either of the quantile
      limits, it will be considered to be in the interval.
}
\author{
Rupert Austin, Aaron Hayman
}
