% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rng_VPC.R
\name{calc_rng_VPC}
\alias{calc_rng_VPC}
\title{Calculation of range VPC.}
\usage{
calc_rng_VPC(exp_data, sim_data, bins, dv_col = "DV", ind_var_col = "TAD",
  exp_quantiles = c(0.05, 0.5, 0.95), prediction_interval = 90,
  min_bin_width = 0.5)
}
\arguments{
\item{exp_data}{data frame containing experimental data.}

\item{sim_data}{data frame containing all simulated data.}

\item{bins}{vector of limits of the time bins.}

\item{dv_col}{name of column containing DV, for both exp_data and sim_data.}

\item{ind_var_col}{name of column containing independent variable (e.g. TAD or TIME), for both exp_data and sim_data.}

\item{exp_quantiles}{vector of required quantiles of experimental DV in each bin.}

\item{prediction_interval}{numeric. Percentage size of prediction interval calculated around all data in each bin of simulated data.}

\item{min_bin_width}{numeric. Width of the smallest possible unrestricted bin (bins will be restricted in width by neighbouring bins). It can be
adjusted to improve the look of the VPC plot.}
}
\value{
A list containing all necessary range VPC parameters to be used in plot_rng_VPC function.
}
\description{
Runs calculations for a more simple version of a VPC where typically, within each bin, the 5th, 50th and 95th percentiles of the experimental
data are calculated along with the central 90\% range of the simulated data (90\% prediction interval). Calculations are run as
non-prediction-corrected and prediction-corrected.
}
\note{
This function has not been designed for bins containing zero points, and may fail or produce strange results if zero points are detected.
}
\author{
Rupert Austin
}
