% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_std_VPC.R
\name{calc_std_VPC}
\alias{calc_std_VPC}
\title{Fast calculation of standard VPC.}
\usage{
calc_std_VPC(exp_data, sim_data, use, bins, dv_col, time_col,
  quantiles = c(0.1, 0.5, 0.9), lower.bound = 0, confidence.interval = 95,
  min_bin_width = 1, T0 = 0)
}
\arguments{
\item{exp_data}{data frame containing experimental data.}

\item{sim_data}{data frame containing all simulated data.}

\item{use}{logic vector of rows to use from the exp_data (must match the sim data too).}

\item{bins}{vector of limits of the time bins.}

\item{dv_col}{name of column containing DV (or IPRED), for both exp_data and sim_data.}

\item{time_col}{name of column containing independent variable (e.g. TAD or TIME), for both exp_data and sim_data.}

\item{quantiles}{vector of quantiles to find between zero and one.}

\item{lower.bound}{numeric. Lower limit of all DV (-Inf will work).}

\item{confidence.interval}{numeric. Percentage size of confidence interval calculated around each quantile in simulated data.}

\item{min_bin_width}{numeric. Width of the smallest possible unrestricted bin (bins will be restricted in width by neighbouring bins).}

\item{T0}{numeric. Start time.}
}
\value{
A list containing all necessary VPC parameters to be used in plot_std_VPC function.
}
\description{
A faster version of previous VPC_data, also allows user defined quantiles.
}
\note{
This function has not been designed for bins containing zero points, and may fail or produce strange results if zero points are detected.
}
\author{
Aaron Hayman
}
