% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColonColon.R
\name{\%::\%}
\alias{\%::\%}
\title{Like R function ' : ' but can be applied to vectors.}
\usage{
x \%::\% y
}
\arguments{
\item{x}{numeric vector.}

\item{y}{numeric vector.}
}
\value{
A numeric vector of the form \deqn{c(x[1]:y[1], x[2]:y[2], ...).}
 If x and y have lengths nx and ny, then x and y are recycled until \deqn{x[nx]:y[ny]} is performed.
}
\description{
x\%::\%y is similar to x:y, except x and y are allowed to be vectors.
}
\examples{
  1\%::\%c(8,7)
  c(1:8,1:7) # are both equal

  c(0,1)\%::\%c(4,7)
  c(0:4,1:7) # are both equal

  c(0,1)\%:\%c(5,4,-3)
  c(0:5,1:4,0:-3,1:5,0:4,1:-3) # are both equal
}
\author{
Aaron Hayman
}
