% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_strat.R
\name{count_strat}
\alias{count_strat}
\title{Function for performing contingency analysis of data counts, with stratification by one or two factors.}
\usage{
count_strat(data, strat_col1, strat_col2 = "", pct = "by_col")
}
\arguments{
\item{data}{data for analysis as a data frame.}

\item{strat_col1}{name of first column for result stratification.}

\item{strat_col2}{name of second column for result stratification. can be set to '' if second stratification is not required.}

\item{pct}{either 'by_col' or 'by_row'. Indicates how count percentages should be calculated.}
}
\value{
Results are returned as a nicely formatted data frame.
}
\description{
Useful for producing stratified sample counts when producing tables for M&S reports.
The function uses the base R table() function followed by lots of reformatting and renaming to produce a more useful output.
}
\author{
Rupert Austin
}
