% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_time_update.R
\name{covariate_time_update}
\alias{covariate_time_update}
\title{Adds column to a NONMEM data file containing values of a time-varying covariate.}
\usage{
covariate_time_update(data_cov, data_dose, data_final, cov_time = "LBDTC",
  dose_time = "EXSTDTC", ID_col = "SID", cov_column = "LBTESTCD",
  cov_name = "ALB", cov_value = "LBSTRESN", date_format_cov = FALSE,
  date_format_dose = TRUE, time_zone = "US/Eastern")
}
\arguments{
\item{data_cov}{data frame containing information about time-varying covariate.}

\item{data_dose}{data frame containing information about when the dose was administered.}

\item{data_final}{data frame to which column with time-varying covariate should be added.}

\item{cov_time}{column name in data_cov containing information about the time when the covariate was measured.}

\item{dose_time}{column name in data_dose containing information about the time when the dose was administered.}

\item{ID_col}{name of the ID column in data_final.}

\item{cov_column}{column name in data_cov containing information about the name of the covariate to be selected.
e.g. weight, albumin, bilirubin etc.}

\item{cov_name}{name of the time-varying covariate.}

\item{cov_value}{name that describes the value (e.g. "Concentration") of the time-varying covariate.}

\item{date_format_cov}{logical. change date format from '\%Y-\%m-\%dT\%H:\%M' to '\%Y-\%m-\%d \%H:\%M'(if TRUE).}

\item{date_format_dose}{logical. change date format from '\%Y-\%m-\%dT\%H:\%M' to '\%Y-\%m-\%d \%H:\%M'(if TRUE).}

\item{time_zone}{name of required timezone. Daylight saving change is applied when time difference is calculated.}
}
\value{
The input data frame with new column added containing values of time-varying covariate.
}
\description{
Adds column to a NONMEM data file containing values of a time-varying covariate.
}
\note{
The data file containing information about the dose records should be filtered to include only records of interest e.g. related
       to a specific drug (e.g. BAY and not co-medication).

Time difference between times when the dose was administered and the covariate was measured is calculated in hours.

The ID columns in the original data files containing information about covariate and dose should be called SUBJIDN.

The time column in data_final (containing dosing and concentration times) should be called TIME.

The format of the dates in the cov_time and dose_time columns should be '\%Y-\%m-\%dT\%H:\%M' or
       '\%Y-\%m-\%d \%H:\%M' (it is assumed that clock time is present).
}
\author{
Katarzyna Nurzynska
}
