% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EqualsEquals.R
\name{\%==\%}
\alias{\%==\%}
\title{For testing equivalence of numbers.}
\usage{
a \%==\% b
}
\arguments{
\item{a}{numeric vector to be tested for equivalence.}

\item{b}{numeric vector to be tested against \eqn{a} for equivalence.}
}
\value{
TRUE if all but the insignificant parts are identical.
}
\description{
Operator that is similar to ==, for comparing two numeric vectors.
}
\details{
This operator works with the option insigbits, which should be an integer between 1 and 8. The option insigbits defines how many bits
         (from least significant) will be considered as insignificant in deciding upon equality.
}
\note{
Default for insigbits is 6, it can be changed by using \deqn{options(insigbits = #),} where \eqn{#} is the desired value.
}
\examples{
a = 90
b = 100*(0.95 - 0.05)  # 90
a == b  # FALSE
a - b  # 1.421085e-14
a \%==\% b  # TRUE
}
\author{
Aaron Hayman
}
