% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unwatermark.R
\name{extract_details}
\alias{extract_details}
\title{find hidden information in a watermarked image}
\usage{
extract_details(p)
}
\arguments{
\item{p}{}
}
\value{
returns a list, first element is TRUE or false indicating if the image checked was a watermark image (with info).
If FALSE then that is all that is in the list.
If TRUE, the second item in the list is a named character vector containing infomation for locating the original file and who made the watermark.
}
\description{
find hidden information in a watermarked image
}
