% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_info}
\alias{get_info}
\title{Gives details on consecutive repeats within a vector.}
\usage{
get_info(col)
}
\arguments{
\item{col}{a vector.}
}
\value{
A data frame with columns 'value', 'num_rep', 'start_index' and 'end_index':

'value' is the elements of the vector excluding any consecutive repeats,

'num_rep' is the number of times each value consecutively occurs,

'start_index' is the start index of each value within the original vector,

'end_index' is the end index of each value within the original vector.
}
\description{
Gives details on consecutive repeats within a vector.
}
\author{
Aaron Hayman
}
