% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INN.R
\name{\%INN\%}
\alias{\%INN\%}
\title{For testing equivalence of numbers contained anywhere within a vector.}
\usage{
a \%INN\% b
}
\arguments{
\item{a}{numeric vector to be tested for equivalence.}

\item{b}{numeric vector to be tested against \eqn{a} for equivalence.}
}
\value{
TRUE for each element of \eqn{b} if all but the insignificant parts are identical to an element in \eqn{a}.
}
\description{
Similar to \%==\% but for testing if numbers in vector \eqn{a} are contained in vector \eqn{b}.
}
\details{
This operator works with the option insigbits, which should be an integer between 1 and 8. The option insigbits defines how many bits
         (from least significant) will be considered as insignificant in deciding upon equality.
}
\note{
Default for insigbits is 6, it can be changed by using \deqn{options(insigbits = #),} where \eqn{#} is the desired value.
}
\examples{
a = c(900,90)
b = c(100*(0.95 - 0.05), 1000*(0.95 - 0.05))  # c(90,900)
a \%in\% b  # FALSE FALSE
a \%==\% b  # FALSE FALSE
a[2] - b[1]  # 1.421085e-14
a[1] - b[2]  # 1.136868e-13
a \%INN\% b  # TRUE TRUE
}
\author{
Aaron Hayman
}
