% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_strsplit.R
\name{multi_strsplit}
\alias{multi_strsplit}
\title{For splitting strings using a vector of several possible splitting points.}
\usage{
multi_strsplit(x, splits)
}
\arguments{
\item{x}{character vector, each element of which is to be split. Other inputs, including a factor, will give an error.}

\item{splits}{character vector containing regular expressions to use for splitting.}
}
\value{
A character vector containing all characters in \eqn{x} that were split using \eqn{splits}.
}
\description{
An adaptation of the strsplit function, which is only able to take a single character string for its split argument (when fixed = TRUE).
}
\note{
Applies fixed = TRUE to underlying strsplit function. See strsplit documentation for explanation of this argument.
}
\examples{
x = c("a+b-c", "d*e/f")
multi_strsplit(x, splits=c('+', '-', '*', '/'))
# [1] "a" "b" "c" "d" "e" "f"
}
\author{
Will Haese-Hill
}
