% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eGOF.R
\name{pIWRES_alt}
\alias{pIWRES_alt}
\title{plots: IWRES VS. IPRED, IWRES VS. ind_var, |IWRES| VS. IPRED, |IWRES| VS. ind_var, QQ plot of IWRES, histogram of IWRES}
\usage{
pIWRES_alt(data, ind_var, log_ind_var, bks, fval, bins_PRED, bins_ind_var,
  dv_units, ind_var_units, loglog, dv_label, pt_col)
}
\arguments{
\item{data}{a data frame}

\item{ind_var}{character, name of the column containing the independent variable}

\item{log_ind_var}{logical. Should the independent variable be plotted on a log scale}

\item{bks}{breaks to be put in a histogram can be single integer, or vector of break points}

\item{fval}{f for lowess smoothing on qq plot}

\item{bins_PRED}{numeric vector bin boundaries for PRED}

\item{bins_ind_var}{numeric vector bin boundaries for independent variable}

\item{dv_units}{character, units of DV}

\item{ind_var_units}{character, units of independent variable}

\item{loglog}{logical whether to plot on log log scale}

\item{dv_label}{character, label for dv in plots}

\item{pt_col}{colour of points}
}
\description{
plots: IWRES VS. IPRED, IWRES VS. ind_var, |IWRES| VS. IPRED, |IWRES| VS. ind_var, QQ plot of IWRES, histogram of IWRES
}
\author{
Rupert Austin
}
