% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eGOF.R
\name{plot_eGOF}
\alias{plot_eGOF}
\title{FUNCTION FOR PRODUCING GOF PLOTS}
\usage{
plot_eGOF(data, type = "CWRES", ind_var = "TAD", log_ind_var = FALSE,
  bks = 30, fval = 0.66, bins_PRED = NA, bins_ind_var = NA,
  dv_units = "ug/L", dv_label = "DV", ind_var_units = "hour",
  loglog = FALSE, pt_col = "gray70", alt_plots = FALSE)
}
\arguments{
\item{data}{Data frame containing output from NONMEM run}

\item{type}{Defines type of plots to return: CWRES / IWRES / NPDE (default='CWRES')}

\item{ind_var}{Name of the independent variable column (default='TAD')}

\item{log_ind_var}{FALSE/TRUE for setting logarithmic x-axis in some of the plots (default=FALSE)}

\item{bks}{Defines number of bins in histograms (default=30)}

\item{fval}{Smoothing setting for LOWESS trend lines (larger values give smore smoothing, default=0.66). Set to NA to exclude trend lines.}

\item{bins_PRED}{Defines bin limits for PRED/IPRED if binned residual display is required  (default=NA)}

\item{bins_ind_var}{Defines bin limits for ind_var if binned residual display is required (default=NA)}

\item{dv_units}{Units of the observations for axis labelling (default='ug/L')}

\item{dv_label}{Axis label for DV in plots (default='DV')}

\item{ind_var_units}{Units of the independent variable for axis labelling (default='hour')}

\item{loglog}{TRUE/FALSE. If TRUE then DV vs PRED and DV vs IPRED are presented on log-log scales, and CWRES/IWRES vs. PRED/IPRED has log x-axis (default=FALSE)}

\item{pt_col}{Colour of points in all plots (default='gray70')}

\item{alt_plots}{Specifies if alternative layout of IWRES plots should be returned FALSE/TRUE (default=FALSE)}
}
\value{
Produces a panel of GOF plots
}
\description{
Function to generate GOF plots for NONMEM models
Similar to the standard GOF plots in BAST1 library except:
    Colours are all grayscale or black
    Units can be added to axis labels,
    LOWESS smoothed lines can be removed (set fval=NA)
    DV, PRED, and IPRED can be set to display on log scale
    Mean residuals in bins can be used instead of LOWESS smooths (set bin limits using bins_PRED and bins_ind_var arguments)
    Alternative layout can be resquested for IWRES plots

Extensive examples of use are available in this location:

\\\\BAST_SERVER\\BAST_Shared_Files\\QA Documents\\BAST Computer Systems Validation\\final_documents\\BAST_NONMEM_USERS_MANUAL\\GOF_VPC_examples\\enhanced_GOF\\run100
}
\details{
If alt_plots=FALSE then output will be:
     CWRES: DV VS. PRED, DV VS. IPRED, CWRES VS. PRED, CWRES VS. ind_var, QQ plot of CWRES, histogram of CWRES
     IWRES: IWRES VS. IPRED, IWRES VS. ind_var, |IWRES| VS. IPRED, |IWRES| VS. ind_var, QQ plot of IWRES, histogram of IWRES
     NPDE:  NPDE VS. ind_var, NPDE VS. PRED, QQ plot of NPDE, histogram of NPDE

     If alt_plots=TRUE then IWRES output will be changed to:
     IWRES: DV VS. PRED, DV VS. IPRED, IWRES VS. PRED, IWRES VS. ind_var, QQ plot of IWRES, histogram of IWRES
}
\author{
Rupert Austin
}
