% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GOF.R
\name{plot_GOF}
\alias{plot_GOF}
\title{Plots GOF involving DV and CWRES}
\usage{
plot_GOF(filepath, data, ind_var, log_ind_var, bks = 20, fval = 0.33,
  name_stub = "")
}
\arguments{
\item{filepath}{character. File path for output GOF plots.}

\item{data}{data frame containing output from NONMEM run.}

\item{ind_var}{name of the independent variable column, usually TIME or TIME_AFTER_DOSE.}

\item{log_ind_var}{logical. If TRUE, plot x-axis on a log scale.}

\item{bks}{integer. Number of breaks for histogram of CWRES.}

\item{fval}{f value in lowess function. Larger values give more smoothness.}

\item{name_stub}{character. Optional string to add to end of image file name.}
}
\value{
A panel of 8 different diagnostic plots.
}
\description{
Generates general GOF plots for NONMEM models.
}
\details{
Uses BAST function GOFplotter.
}
\author{
Rupert Austin
}
