% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GOF_multi.R
\name{plot_GOF_multi}
\alias{plot_GOF_multi}
\title{Plots multiple GOF involving DV and CWRES}
\usage{
plot_GOF_multi(mod_folder, ind_var = "TIME", log_ind_var = FALSE,
  all = FALSE)
}
\arguments{
\item{mod_folder}{folder path containing the model run folders.}

\item{ind_var}{name of the independent variable column, usually TIME or TIME_AFTER_DOSE.}

\item{log_ind_var}{logical. If TRUE, plot x-axis on a log scale.}

\item{all}{logical. If TRUE, produce GOF plots for each run folder in mod_folder (and not just those where no GOF is present).}
}
\value{
A panel of 8 different diagnostic plots for each run folder in mod_folder.
}
\description{
Generates GOF plots involving DV and CWRES for all run folders within a modelling folder.
}
\details{
Uses BAST function plot_GOF.
}
\note{
Assumes all model folders are in format: "run---".
}
\author{
Will Haese-Hill
}
