% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_IWRES.R
\name{plot_IWRES}
\alias{plot_IWRES}
\title{Plots GOF involving IWRES}
\usage{
plot_IWRES(filepath, data, ind_var, log_ind_var, bks = 20, fval = 0.33,
  name_stub = "")
}
\arguments{
\item{filepath}{character. File path for output IWRES plots.}

\item{data}{data frame containing output from NONMEM run.}

\item{ind_var}{name of the independent variable column, usually TIME or TIME_AFTER_DOSE.}

\item{log_ind_var}{logical. If TRUE, plot x-axis on a log scale.}

\item{bks}{integer. Number of breaks for histogram of IWRES.}

\item{fval}{f value in lowess function. Larger values give more smoothness.}

\item{name_stub}{character. Optional string to add to end of image file name.}
}
\value{
A panel of 4 different diagnostic IWRES plots.
}
\description{
Generates general IWRES plots for NONMEM models.
}
\details{
Uses BAST function IWRESplotter.
}
\author{
Rupert Austin
}
