% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rng_VPC.R
\name{plot_rng_VPC}
\alias{plot_rng_VPC}
\title{Plot range VPC using output of calc_rng_VPC function.}
\usage{
plot_rng_VPC(VPC_results, x_lab = "Time after dose (hours)",
  y_lab = "Prediction-corrected concentration (ug/L)",
  add_simulated_median = TRUE, prediction_corrected = TRUE,
  plot_obs = TRUE, bw = FALSE, legend_position = "topright",
  legend_size = 0.7, col_scheme = FALSE, ...)
}
\arguments{
\item{VPC_results}{list. Output of calc_rng_VPC function.}

\item{x_lab}{label for the x axis.}

\item{y_lab}{label for the y axis.}

\item{add_simulated_median}{TRUE (default) or FALSE. Add simulated median quantile to bins in plot.}

\item{prediction_corrected}{TRUE (default) or FALSE. Produce a prediction-corrected range VPC plot.}

\item{plot_obs}{TRUE (default) or FALSE. Overlay observed data points on plot.}

\item{bw}{TRUE or FALSE (default). If set to TRUE a black and white output is given, which is required for certain submission reports.}

\item{legend_position}{FALSE or "bottomright"/"bottom"/"bottomleft"/"left"/"topleft"/"top"/"topright"/"right"/"center". Descriptive position of legend in plot.
FALSE ensures no legend is produced.}

\item{legend_size}{positive numeric. Typically in range 0.4 to 3 (from cex argument in legend).}

\item{col_scheme}{defaults to FALSE. Can be used for setting a user-defined colouring of VPC plot elements using the format of: \deqn{col_scheme = c('gray90', 'black', 'orange', 'red'),} where:

       1st colour is for shading of prediction interval of simulated data.

       2nd colour is for 50th percentile of simulated data.

       3rd colour is for outer experimental percentiles.

       4th colour is for 50th experimental percentile.}
}
\value{
A range VPC plot. Lines showing percentiles of experimental data in each bin are shown. A prediction interval around all simulated data
        in each bin is also shown.
}
\description{
Function for production of range VPC plots.
}
\note{
The user can add standard R plot() arguments such as log='y', xlim=c(0,100), yaxt='n', main='VPC' etc. to customize the VPC plot.
}
\author{
Rupert Austin
}
