% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eGOF.R
\name{pNPDE}
\alias{pNPDE}
\title{plots: NPDE VS. ind_var, NPDE VS. PRED, QQ plot of NPDE, histogram of NPDE}
\usage{
pNPDE(data, ind_var, log_ind_var, bks, fval, bins_PRED, bins_ind_var, dv_units,
  ind_var_units, pt_col)
}
\arguments{
\item{data}{a data frame}

\item{ind_var}{character, name of the column containing the independent variable}

\item{log_ind_var}{logical. Should the independent variable be plotted on a log scale}

\item{bks}{breaks to be put in a histogram can be single integer, or vector of break points}

\item{fval}{f for lowess smoothing on qq plot}

\item{bins_PRED}{numeric vector bin boundaries for PRED}

\item{bins_ind_var}{numeric vector bin boundaries for independent variable}

\item{dv_units}{character, units of DV}

\item{ind_var_units}{character, units of independent variable}

\item{pt_col}{colour of points}
}
\description{
plots: NPDE VS. ind_var, NPDE VS. PRED, QQ plot of NPDE, histogram of NPDE
}
\author{
Rupert Austin
}
