% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabRsrc.R
\name{read_TAB}
\alias{read_TAB}
\title{Fast reading of NONMEM $TABLE files.}
\usage{
read_TAB(p, headers = TRUE)
}
\arguments{
\item{p}{a character string containing the path of the .tab file.}

\item{headers}{logical. Whether headers should be included in output data frame.}
}
\value{
A data frame containing the .tab file data.
}
\description{
Fast reading of NONMEM $TABLE files by utilising Rcpp package. Replaces read.table function.
}
\examples{
\donttest{read_TAB(p = "...\\\\data.tab",headers = F)}
}
\author{
Aaron Hayman
}
