% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_col.R
\name{set_col}
\alias{set_col}
\title{For setting colours and associated densities.}
\usage{
set_col(col, den = 255)
}
\arguments{
\item{col}{character. Name of colour.}

\item{den}{numeric. Density range from 0 (transparent) to 255 (opaque).}
}
\value{
A colour hex code, for use in plotting etc.
}
\description{
Specify a colour by name, and alter the density without having to convert to RGB.
}
\note{
col can have length greater than 1, and the function will output a character of the same length.
}
\references{
For a list of named colours recognised by R, see: \url{http://www.stat.columbia.edu/~tzheng/files/Rcolor.pdf}.
}
\author{
Will Haese-Hill
}
