% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{strat_stats}
\alias{strat_stats}
\title{Get summary statistics on a vector of numbers, stratified by levels of a vector of factors.}
\usage{
strat_stats(id, da, quantiles = NULL)
}
\arguments{
\item{id}{vector of any type (except factor) to describe the various levels, with same length as da.}

\item{da}{numeric vector of data.}

\item{quantiles}{numeric. Between 0 and 1 to determine 0\% - 100\% quantile for each id.}
}
\value{
A data frame with a row for each level (see Details).
}
\description{
Gives basic statistics for a data item da (ignores NAs), within different levels id. \cr
WARNING - If data is very large, function can encounter serious rounding issues!
}
\details{
Returned data frame contains the following:
\describe{
   \item{id}{level identifier.}
   \item{n}{number of items in level.}
   \item{min}{minimum value of da in level.}
   \item{max}{maximum value of da in level.}
   \item{sum}{sum of values of da in level.}
   \item{mean}{mean value of da in level.}
   \item{median}{median value of da in level.}
   \item{SD}{standard deviation value of da in level.}
   \item{pct_n\%}{specified quantiles of levels.}
 }
}
\examples{
 # Want to compare weights of different animals from the following data frame:

 # ANIMAL   WEIGHT
 # Chicken  3.0
 # Cat	    4.2
 # Cat	    3.1
 # Dog	    12.8
 # Cat	    6.1
 # Dog	    20.4
 # Cat	    2.3

 id = c("Chicken", "Cat", "Cat", "Dog", "Cat", "Dog", "Cat")
 da = c(3.0, 4.2, 3.1, 12.8, 6.1, 20.4, 2.3)
 strat_stats(id,da)

 #        id n  min  max  sum   mean median       SD
 # 1     Cat 4  2.3  6.1 15.7  3.925   3.65 1.645955
 # 2 Chicken 1  3.0  3.0  3.0  3.000   3.00      NaN
 # 3     Dog 2 12.8 20.4 33.2 16.600  16.60 5.374012
}
\author{
Aaron Hayman
}
