% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_stats.R
\name{sum_stats}
\alias{sum_stats}
\title{For calculating summary statistics on a vector.}
\usage{
sum_stats(data, pct = c(5, 95))
}
\arguments{
\item{data}{numeric vector to compute summary statistics on.}

\item{pct}{defaults to \eqn{c(5, 95)}. Up to 3 different percentiles can be calculated. For example: \deqn{sum_stats(data, pct = c(5, 50, 95)).}
If percentiles are not required then set: \deqn{pct = NA.}}
}
\value{
A vector of named elements (see Details).
}
\description{
Calculates arithmetic and geometric summary statistics for an input vector.
}
\details{
Vector of summary statistics contains the following:
         \describe{
           \item{n}{number of elements.}
           \item{n_NA}{number of NA elements.}
           \item{min}{minimum value.}
           \item{max}{maximum value.}
           \item{median}{median value.}
           \item{arith_mean}{arithmetic mean value.}
           \item{std_dev}{standard deviation.}
           \item{arith_CV}{arithmetic coefficient of variation.}
           \item{geo_mean}{geometric mean value.}
           \item{geo_std_dev}{geometric standard deviation.}
           \item{geo_CV}{geometric coefficient of variation.}
         }
}
\author{
Rupert Austin
}
