% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unwatermark.R
\name{unwatermark}
\alias{unwatermark}
\title{replace a watermarked image with original}
\usage{
unwatermark(path, folder = TRUE)
}
\arguments{
\item{path}{character, path to a file or folder containing watermaked images}

\item{folder}{Logical, whether the path is a folder or a file}
}
\value{
no return, the function will replace watermarked images with originals when and if it finds them
}
\description{
replace a watermarked image with original
}
