% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v.R
\name{v}
\alias{v}
\title{Invoke a Data Viewer.}
\usage{
v(x, title)
}
\arguments{
\item{x}{an R object which can be coerced to a data frame with non-zero numbers of rows and columns.}

\item{title}{title for viewer window. Defaults to name of x prefixed by Data:.}
}
\value{
Invisible NULL. The function puts up a window and returns immediately: the window can be closed via its controls or menus.
}
\description{
In RStudio sometimes you want to see all the data, but View is introduced in a different namespace.
}
\author{
Aaron Hayman
}
