% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watermark.R
\name{watermark}
\alias{watermark}
\title{Create a copy of a plot with a watermark}
\usage{
watermark(path, text = "Preliminary for internal discussion",
  colour = "#80808030", prefix = "W_", suffix = "", scale = 1)
}
\arguments{
\item{path}{file path of the plot image to have a watermark added to.}

\item{text}{text to display as the watermark.  Default is 'Preliminary for internal discussion.}

\item{colour}{colour to be used for watermark, can be partially transparent. Default is '#80808030', a translucent grey.}

\item{prefix}{A string to add to the start of the filename. Default is "W_".}

\item{suffix}{A string to add to the end of the filename. The default is empty, this may be used as an alternative to the prefix.}

\item{scale}{A number to rescale watermark. Default is one, which should take up as much space as possible.}
}
\description{
To avoid errors caused by mistakenly using graphs not intended publication, a copy of graphs with a watermark added should
be used in draft publications. This function is designed to make a copy of an image and add a watermark to it. The source can be any image
file accepted by \link[magick]{image_read}. The function will produce a .png file in the same location with filename identical accept for
addition of a prefix/suffix.
}
\examples{
{
\dontrun{

files = dir()
files=files[grep('f_.+[.]png$',files)]
for(f in files) watermark(f)
}
}

}
\author{
Blesson Chacko and Aaron Hayman.
}
